#53756 Ensure nginx module configures IPv6 addresses even though facter does not find any IPv6 address
IP addresses will be configured by puppet, but facter data is fetched before the puppet run. This creates a situation where IPv6 will be actually configured, but the nginx module does not realize it and skips the IPv6 configuration.

This change, based on changes in 0.8.0 of the puppet-nginx module, changes the related checks.